Refactored modules and moved training to train.py
